requirements.txt
argparse==1.1
torch==1.8.1
torchvision==0.9.1
numpy==1.14.3
imageio==2.3.0
opencv-python==4.1.2
requests==2.27.1
scikit-image==0.13.1
PyYAML==3.12
einops==0.4.0
six==1.11.0
tqdm==4.62.3
Pillow==5.1.0
scipy==1.1.0
matplotlib==2.2.2
scikit-image==0.17.2
scikit-learn==0.19.1
scipy=1.1.0
PyWavelets==1.1.1
requests==2.26.0
networkx==2.1
imageio==2.3.0
h5py==2.7.1
setuptools==57.4.0
tifffile==2020.9.3
cryptography==2.2.2
tensorboardX
